creat new main section
