+++ front/src/pages/index.astro
@@ -1,6 +1,5 @@
 ---
 import Layout from "../layouts/Layout.astro"
-import { Main } from "../widgets/Main/Main"
 import { Portfolio } from "../widgets/Portfolio/Portfolio"
 import { Steps } from "../widgets/Steps/Steps"
 import { Services } from "../widgets/Services/Services"
@@ -8,10 +7,11 @@ import { Artists } from "../widgets/Artists/Artists"
 import { Testimonials } from "../widgets/Testimonials/Testimonials"
 import { Faq } from "../widgets/Faq/Faq"
 import { FormSection } from "../widgets/FormSection/FormSection"
+import { Main } from "../widgets/Main/Main"
 ---
 
 <Layout>
-    <Main client:load />
+    <Main />
     <Portfolio client:only="preact" />
     <Steps />
     <Services client:load />
