+++ front/src/widgets/Main/Main.module.scss
@@ -1,85 +1,40 @@
 .wrapper {
-    position: relative;
-
-    &::after {
-        content: "";
-        position: absolute;
-        z-index: 1;
-        width: 100%;
-        height: calc(var(--index) * 10);
-        background-image: url(./const/ground.png);
-        background-size: cover;
-        background-repeat: no-repeat;
-        background-position: center;
-        bottom: calc(var(--index) * -4.5);
-    }
-}
-
-.container {
-    height: 100vh;
     display: flex;
     align-items: center;
     justify-content: center;
-    text-align: center;
-    overflow: hidden;
-    position: relative;
-}
-
-.layer {
-    height: 100%;
     width: 100%;
-    position: absolute;
+    height: 100vh;
+    padding: 0 10px;
+    background-image: url("./const/background.jpg");
+    background-repeat: no-repeat;
     background-size: cover;
-    background-position: center;
-    will-change: transform;
-    transition: var(--transition);
-    z-index: 2;
-}
 
-.layerBase {
-    transform: translate3d(0, calc(var(--scrollTop) / 1.6), 0);
-    z-index: 0;
-}
+    .container {
+        margin-top: 100px;
+        position: relative;
+        width: fit-content;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        gap: 10px;
 
-.layerMiddle {
-    transform: translate3d(0, calc(var(--scrollTop) / 2.5), 0);
-}
+        .title {
+            position: relative;
+            width: 550px;
+            z-index: 1;
+        }
 
-.layerFront {
-    transform: translate3d(0, calc(var(--scrollTop) / 5.7), 0);
-}
+        .caption {
+            position: relative;
+            width: 320px;
+            z-index: 1;
+        }
 
-.textContainer {
-    z-index: 1;
-    transform: translate3d(0, calc(var(--scrollTop) / 2), 0);
-    transition: var(--transition);
-    will-change: transform;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 20px;
-    margin: -40px 20px 0;
-    position: relative;
-}
-
-.title {
-    width: 550px;
-}
-
-.caption {
-    width: 320px;
-}
-
-.horns {
-    position: absolute;
-    top: -170px;
-    z-index: -1;
-    transform: translate3d(0, calc(var(--scrollTop) / 6), 0);
-}
-
-.scroll {
-    z-index: 2;
-    position: absolute;
-    bottom: 50px;
-    cursor: pointer;
+        .horns {
+            position: absolute;
+            top: -170px;
+            left: 50%;
+            transform: translateX(-50%);
+        }
+    }
 }
