+++ front/src/widgets/MainParallax/const/layer-base.png
