+++ front/src/widgets/MainParallax/const/layer-front.png
