+++ front/src/widgets/MainParallax/const/layer-middle.png
